댓글 수정, 삭제시 별점 테이블 반영

+++ src/main/java/com/ridibooks/clone_ridibooks_be/dto/StarsRequestDto.java
@@ -4,7 +4,7 @@ import lombok.Getter;
 
 @Getter
 public class StarsRequestDto {
-    private Float avgStar;
+    private Double avgStar;
     private Long totalCount;
     private int star5Count;
     private int star4Count;
